Calculate & export performance metrics for integrated samples - so far only for specific depths

This function integrates a selected variable (e.g., DO, temperature) across specified depths for each model output, compares it to observed data, and returns summary statistics.

The aim of the function is to integrate the model outputs across specified depth intervals since observed water sample data are often collected as integrated samples over multiple depths. There are two options in this function: 1)The user either specify an exact depth 2)The user give provides a depth range as a vector (e.g., `c(lower, upper)`) 3)The user gives dataframe with dynamic lower-upper bounds for certain period. The dataframe should have columns datetime, lower, upper the function integrate the average values within these depts.

Usage

export_all_stats(output, variable, depths, bathy_path, obs_data)

integrate_depths(
  output,
  depth,
  variable,
  bathy_path,
  obs_data = NULL,
  folder = NULL,
  hypso_avg = TRUE,
  dz = 0.1
)

Arguments

output

List: A list containing model output data for different models at different depths.

variable

character: variable name to be extracted

depths

Numeric vector: One or more depths to extract or integrate over. But so far it just works with single vector output

bathy_path

character: variable name to be extracted

obs_data

list: list of observed data

depth

numeric or vector or dataframe: 1)if exact depth is being extracted, provide a simple numeric value, 2)if integrating across depths for all time series, provide a vector c(lower, upper) 3) if you want to integrate different depth intervals for different dates, a dataframe with datetime, lower, upper values should be given

folder

character: main filepath; where all the output is stored

hypso_avg

logical: if TRUE= hypsometric curve (depth-area relationship) will be used, otherwise average of selected columns will be calculated

dz

numeric: interpolation depth interval. The default is 0.1

Value

A data frame of calculated statistics (e.g., RMSE, KGE) for each model, depth, and variable.

A list of integrated variables for the specified depth or depth range.